+++ src/Products/PloneMeeting/tests/helpers.py
@@ -166,70 +166,93 @@ class PloneMeetingTestingHelpers:
     def get_transitions_for_proposing_item(self, first_level=False):
         meetingConfigNumber = self._determinateUsedMeetingConfigNumber()
         if first_level:
-           return getattr(self,
-                   ('TRANSITIONS_FOR_PROPOSING_ITEM_FIRST_LEVEL_%d' % meetingConfigNumber))
+            return getattr(
+                self, ('TRANSITIONS_FOR_PROPOSING_ITEM_FIRST_LEVEL_%d' % meetingConfigNumber))
         else:
-           return getattr(self,
-                   ('TRANSITIONS_FOR_PROPOSING_ITEM_%d' % meetingConfigNumber))
+            return getattr(
+                self, ('TRANSITIONS_FOR_PROPOSING_ITEM_%d' % meetingConfigNumber))
 
-    def proposeItem(self, item, first_level=False):
+    def proposeItem(self, item, first_level=False, as_manager=True):
         '''Propose passed p_item using TRANSITIONS_FOR_PROPOSING_ITEM_x.
            The p_meetingConfigNumber specify if we use meetingConfig or meetingConfig2, so
            the _x here above in TRANSITIONS_FOR_PROPOSING_ITEM_x is 1 or 2.
            If p_first_level is True, we will use TRANSITIONS_FOR_PROPOSING_ITEM_FIRST_LEVEL_x,
            this makes it possible to reach an intermediate propose level.'''
-        self._doTransitionsFor(item, self.get_transitions_for_proposing_item(first_level))
+        self._doTransitionsFor(item,
+                               self.get_transitions_for_proposing_item(first_level),
+                               as_manager=as_manager)
 
-    def prevalidateItem(self, item):
+    def prevalidateItem(self, item, as_manager=True):
         '''Prevalidate passed p_item using TRANSITIONS_FOR_PREVALIDATING_ITEM_x.
            The p_meetingConfigNumber specify if we use meetingConfig or meetingConfig2, so
            the _x here above in TRANSITIONS_FOR_PREVALIDATING_ITEM_x is 1 or 2.'''
         meetingConfigNumber = self._determinateUsedMeetingConfigNumber()
-        self._doTransitionsFor(item, getattr(self, ('TRANSITIONS_FOR_PREVALIDATING_ITEM_%d' % meetingConfigNumber)))
+        self._doTransitionsFor(
+            item,
+            getattr(self, ('TRANSITIONS_FOR_PREVALIDATING_ITEM_%d' % meetingConfigNumber)),
+            as_manager=as_manager)
 
-    def validateItem(self, item):
+    def validateItem(self, item, as_manager=True):
         '''Validate passed p_item using TRANSITIONS_FOR_VALIDATING_ITEM_x.
            The p_meetingConfigNumber specify if we use meetingConfig or meetingConfig2, so
            the _x here above in TRANSITIONS_FOR_VALIDATING_ITEM_x is 1 or 2.'''
         meetingConfigNumber = self._determinateUsedMeetingConfigNumber()
-        self._doTransitionsFor(item, getattr(self, ('TRANSITIONS_FOR_VALIDATING_ITEM_%d' % meetingConfigNumber)))
+        self._doTransitionsFor(
+            item,
+            getattr(self, ('TRANSITIONS_FOR_VALIDATING_ITEM_%d' % meetingConfigNumber)),
+            as_manager=as_manager)
 
-    def presentItem(self, item):
+    def presentItem(self, item, as_manager=True):
         '''Present passed p_item using TRANSITIONS_FOR_PRESENTING_ITEM_x.
            The p_meetingConfigNumber specify if we use meetingConfig or meetingConfig2, so
            the _x here above in TRANSITIONS_FOR_PRESENTING_ITEM_x is 1 or 2.'''
         meetingConfigNumber = self._determinateUsedMeetingConfigNumber()
-        self._doTransitionsFor(item, getattr(self, ('TRANSITIONS_FOR_PRESENTING_ITEM_%d' % meetingConfigNumber)))
+        self._doTransitionsFor(
+            item,
+            getattr(self, ('TRANSITIONS_FOR_PRESENTING_ITEM_%d' % meetingConfigNumber)),
+            as_manager=as_manager)
 
-    def publishMeeting(self, meeting):
+    def publishMeeting(self, meeting, as_manager=False):
         '''Publish passed p_meeting using TRANSITIONS_FOR_PUBLISHING_MEETING_x.
            The p_meetingConfigNumber specify if we use meetingConfig or meetingConfig2, so
            the _x here above in TRANSITIONS_FOR_PUBLISHING_MEETING_x is 1 or 2.'''
         meetingConfigNumber = self._determinateUsedMeetingConfigNumber()
-        self._doTransitionsFor(meeting, getattr(self, ('TRANSITIONS_FOR_PUBLISHING_MEETING_%d' % meetingConfigNumber)))
+        self._doTransitionsFor(
+            meeting,
+            getattr(self, ('TRANSITIONS_FOR_PUBLISHING_MEETING_%d' % meetingConfigNumber)),
+            as_manager=as_manager)
 
-    def freezeMeeting(self, meeting):
+    def freezeMeeting(self, meeting, as_manager=False):
         '''Freeze passed p_meeting using TRANSITIONS_FOR_FREEZING_MEETING_x.
            The p_meetingConfigNumber specify if we use meetingConfig or meetingConfig2, so
            the _x here above in TRANSITIONS_FOR_FREEZING_MEETING_x is 1 or 2.'''
         meetingConfigNumber = self._determinateUsedMeetingConfigNumber()
-        self._doTransitionsFor(meeting, getattr(self, ('TRANSITIONS_FOR_FREEZING_MEETING_%d' % meetingConfigNumber)))
+        self._doTransitionsFor(
+            meeting,
+            getattr(self, ('TRANSITIONS_FOR_FREEZING_MEETING_%d' % meetingConfigNumber)),
+            as_manager=as_manager)
 
-    def decideMeeting(self, meeting):
+    def decideMeeting(self, meeting, as_manager=False):
         '''Decide passed p_meeting using TRANSITIONS_FOR_DECIDING_MEETING_x.
            The p_meetingConfigNumber specify if we use meetingConfig or meetingConfig2, so
            the _x here above in TRANSITIONS_FOR_DECIDING_MEETING_x is 1 or 2.'''
         meetingConfigNumber = self._determinateUsedMeetingConfigNumber()
-        self._doTransitionsFor(meeting, getattr(self, ('TRANSITIONS_FOR_DECIDING_MEETING_%d' % meetingConfigNumber)))
+        self._doTransitionsFor(
+            meeting,
+            getattr(self, ('TRANSITIONS_FOR_DECIDING_MEETING_%d' % meetingConfigNumber)),
+            as_manager=as_manager)
 
-    def closeMeeting(self, meeting):
+    def closeMeeting(self, meeting, as_manager=False):
         '''Close passed p_meeting using TRANSITIONS_FOR_CLOSING_MEETING_x.
            The p_meetingConfigNumber specify if we use meetingConfig or meetingConfig2, so
            the _x here above in TRANSITIONS_FOR_CLOSING_MEETING_x is 1 or 2.'''
         meetingConfigNumber = self._determinateUsedMeetingConfigNumber()
-        self._doTransitionsFor(meeting, getattr(self, ('TRANSITIONS_FOR_CLOSING_MEETING_%d' % meetingConfigNumber)))
+        self._doTransitionsFor(
+            meeting,
+            getattr(self, ('TRANSITIONS_FOR_CLOSING_MEETING_%d' % meetingConfigNumber)),
+            as_manager=as_manager)
 
-    def backToState(self, itemOrMeeting, state):
+    def backToState(self, itemOrMeeting, state, as_manager=False):
         """Set the p_item back to p_state.
            Given p_state MUST BE original state name, aka state existing in PloneMeeting workflow."""
         # if a wf path is defined in BACK_TO_WF_PATH_x to go to relevant state, use it
@@ -247,7 +270,7 @@ class PloneMeetingTestingHelpers:
         self.changeUser('admin')
         max_attempts = 20
         nb_attempts = 0
-        while not itemOrMeeting.queryState() == state or nb_attempts >= max_attempts:
+        while not itemOrMeeting.queryState() == state and nb_attempts <= max_attempts:
             nb_attempts += 1
             if not useDefinedWfPath:
                 transitions = self.transitions(itemOrMeeting)
@@ -259,15 +282,17 @@ class PloneMeetingTestingHelpers:
             raise ValueError('impossible to go back to {}'.format(state))
         self.changeUser(currentUser)
 
-    def _doTransitionsFor(self, itemOrMeeting, transitions):
+    def _doTransitionsFor(self, itemOrMeeting, transitions, as_manager=False):
         """Helper that just trigger given p_transitions on given p_itemOrMeeting."""
         # do things as admin to avoid permission issues
-        currentUser = self.member.getId()
-        self.changeUser('admin')
+        if as_manager:
+            currentUser = self.member.getId()
+            self.changeUser('admin')
         for tr in transitions:
             if tr in self.transitions(itemOrMeeting):
                 self.do(itemOrMeeting, tr)
-        self.changeUser(currentUser)
+        if as_manager:
+            self.changeUser(currentUser)
 
     def _determinateUsedMeetingConfigNumber(self):
         """Helper method that check if we use meetingConfig or meetingConfig2."""
@@ -306,7 +331,8 @@ class PloneMeetingTestingHelpers:
         for extra_user_id in user_ids:
             user = api.user.get(extra_user_id)
             # remove from every groups, bypass Plone groups (including virtual)
-            for group_id in [user_group_id for user_group_id in user.getGroups() if '_' in user_group_id]:
+            for group_id in [user_group_id for user_group_id in user.getGroups()
+                             if '_' in user_group_id]:
                 api.group.remove_user(groupname=group_id, username=extra_user_id)
         cleanRamCacheFor('Products.PloneMeeting.ToolPloneMeeting._users_groups_value')