+++ CHANGES.rst
@@ -5,8 +5,25 @@ Changelog
 4.2a6 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fix advanced search view with collective.solr
+- Small fixes in the test to improve MeetingLalouviere test run
+- Fixed a misstyped condition in tests/helpers.py
+- Added new type of presence for item attendee (used to ignore an attendee on some items) :
 
+  - new meeting optional attribute `non attendee`;
+  - may be used in addition to `present/absent/excused` as even an absent attendee may be set non attendee for a specific item
+  - changed parameter `patterns` on `print_in_and_out_attendees` to `custom_patterns` to be able to redefine only one single pattern
+- Fixed `AskedAdvicesVocabulary` ram.cache cachekey to avoid same vocabulary used for 2 different MeetingConfigs
+  (the `indexAdvisers` term on DashboardCollection was using another MeetingConfig values), moreover made it more robust in case weird context is received
+- Execute the `MeetingConfig.onMeetingTransitionItemActionToExecute` TAL expressions as `Manager` in `utils.meetingExecuteActionOnLinkedItems`
+  to avoid permission problems, what is defined in the configuration must be applied.
+  This makes the `a power observer may only access accepted items when meeting is closed` work when current user is a `MeetingManager`,
+  not a `Manager`, instead having a permission error as `MeetingItem.updateLocalRoles` is protected with the `Modify portal content` permission
+- Added new optional field (decisionSuite) for item
+- In tests WF helpers (validateItem, decideMeeting, ...) added parameter as_manager, True by default for MeetingItem related methods and False
+  by default for Meeting related methods.  This way we avoid as much as possible hidden permission problems
+- Exclude SearchableText indexing for IAnnex objects
+- Make sure CKeditor panels are dispayed correctly in popups (adding/editing advice)
 
 4.2a5 (2020-03-17)
 ------------------